pyproject.toml
[project]
name = "iasa-asp"
version = "0.1.0"
description = "Audio Signal Processing course at IASA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kaggle==1.6.3",
    "librosa==0.9.2",
    "notebook==7.3.2",
    "numpy==1.23.4",
    "pre-commit==4.1.0",
    "pyannote-audio==3.3.2",
    "scipy==1.9.3",
    "soxr==0.5.0",
    "torch==2.5.0",
    "torchaudio==2.5.0",
    "webrtcvad==2.0.10",
    "zff-vad",
]

[tool.uv.sources]
zff-vad = { git = "https://github.com/idiap/zff_vad.git" }
